Added phrase creawtion UI, some base phrase logic. Need to write/test creat phrase wallet async code

+++ app/src/main/kotlin/com/caplaninnovations/looprwallet/handlers/PermissionHandler.kt
@@ -38,8 +38,8 @@ open class PermissionHandler(private val activity: BaseActivity,
         ActivityCompat.OnRequestPermissionsResultCallback {
 
     @IntDef(
-            REQUEST_CODE_CAMERA.toLong(),
-            REQUEST_CODE_EXTERNAL_FILES.toLong()
+            REQUEST_CODE_CAMERA,
+            REQUEST_CODE_EXTERNAL_FILES
     )
     annotation class Code
